Move transaction_types to wasm_bindings
