+++ quadratic-core/src/grid/sheet/rendering.rs
@@ -317,12 +317,13 @@ impl Sheet {
 #[cfg(test)]
 mod tests {
     use crate::{
-        controller::{transaction_types::JsCodeResult, GridController},
+        controller::GridController,
         grid::{
             js_types::{JsHtmlOutput, JsRenderCell},
             Bold, CellAlign, CodeCellLanguage, CodeCellRunOutput, CodeCellRunResult, CodeCellValue,
             Italic, RenderSize,
         },
+        wasm_bindings::transaction_types::JsCodeResult,
         CellValue, Error, ErrorMsg, Pos, Rect, Value,
     };
